fix: Ensuring ProjectId is represented as a long datatype instead of int

+++ src/Atc.Kepware.Configuration.Contracts/Interfaces/IotGateway/IIotAgentBase.cs
@@ -5,7 +5,7 @@ public interface IIotAgentBase
     /// <summary>
     /// The ProjectId.
     /// </summary>
-    int ProjectId { get; set; }
+    long ProjectId { get; set; }
 
     /// <summary>
     /// Name of the Iot Agent.
